review(chart): apply Task 5 reviewer findings
- Split PositionsChartRenderRow.swift into three files
  (RenderRow / Mode / BaselineResolver) so each top-level type
  has its own file (CODE_GUIDE.md §2 — one primary type per file).
- Mark legend helpers (PositionsChartLegendItem, DashedLineSwatch,
  ProfitLossLegendSwatch) `private` so they don't leak module-wide.
  Renames ProfitLossPositionsChartLegendItem → ProfitLossLegendSwatch
  (the over-eager find-and-replace from the original commit).
- Add reason comment to PositionsChart.swift's file-top
  swiftlint:disable directive (CODE_GUIDE.md §3).
- Append "dashed line" / "solid line" to the legend item's VoiceOver
  label so it disambiguates two entries that differ only in line
  style (UI_GUIDE.md §5 — color is never the sole channel).
